share test utils and add unit to package.json
